+++ galsim/zernike.py
@@ -427,9 +427,9 @@ class Zernike(object):
     @param R_inner  Inner radius.  [default: 0.0]
     """
     def __init__(self, coef, R_outer=1.0, R_inner=0.0):
-        self.coef = np.asarray(coef)
+        self.coef = np.asarray(coef, dtype=float)
         if len(self.coef) <= 1:
-            self.coef = np.array([0, 0])
+            self.coef = np.array([0, 0], dtype=float)
         self.R_outer = float(R_outer)
         self.R_inner = float(R_inner)
